requirements.txt
asgiref==3.7.2
Django==3.2.22
django-phonenumber-field==7.2.0
phonenumbers==8.13.23
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
defusedxml==0.8.0rc2
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
idna==3.4
install==1.3.5
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.4.2
urllib3==2.0.7
